+++ resources/js/components/Listing/Listing.vue
@@ -33,6 +33,10 @@ export default {
             type: Boolean,
             default: false,
         },
+        transformItems: {
+            type: Function,
+            default: (items) => items,
+        },
     },
 
     data: () => ({
@@ -55,12 +59,14 @@ export default {
         },
 
         hitsPerPage() {
+            let hasDefault = this.$root.config.grid_per_page_values.includes(this.$root.config.grid_per_page)
+
             return this.$root.config.grid_per_page_values
                 .map(function (pages, index) {
                     return {
                         label: pages,
                         value: pages,
-                        default: pages == config.grid_per_page,
+                        default: hasDefault ? pages == config.grid_per_page : index == 0,
                     }
                 })
                 .concat({ label: this.$root.config.translations.all, value: 10000 })
@@ -165,7 +171,7 @@ export default {
             let data = uiState[this.index]
 
             // Remove the root path from the category if it's in there
-            let category = data.hierarchicalMenu?.category_lvl1
+            let category = [...(data.hierarchicalMenu?.category_lvl1 ?? [])]
             for (let i = 0; i < this.rootPath?.length && category?.length && category[0] == this.rootPath[i]; i++) {
                 category.splice(0, 1)
             }
@@ -174,7 +180,7 @@ export default {
                 ...(data.range || {}),
                 ...(data.refinementList || {}),
                 category: category?.length ? category.join('--') : undefined,
-                q: data.query,
+                q: data.query !== '__NO_QUERY__' ? data.query : undefined,
                 page: data.page > 0 ? String(data.page) : undefined,
                 sort: data.sortBy,
                 hits: data.hitsPerPage != config.grid_per_page ? data.hitsPerPage : undefined,
@@ -185,9 +191,9 @@ export default {
             let ranges = Object.fromEntries(Object.entries(routeState).filter(([key]) => this.rangeAttributes.includes(key)))
 
             let refinementList = Object.fromEntries(
-                Object.entries(routeState).filter(
-                    ([key]) => !['q', 'hits', 'sort', 'page', 'category'].includes(key) && !this.rangeAttributes.includes(key),
-                ),
+                Object.entries(routeState)
+                    .filter(([key]) => !['q', 'hits', 'sort', 'page', 'category'].includes(key) && !this.rangeAttributes.includes(key))
+                    .map(([key, refinement]) => [key, typeof refinement === 'string' ? refinement.split(',') : refinement]),
             )
 
             const categories = [...(this.rootPath || []), ...(routeState.category?.split('--') || [])]
@@ -197,7 +203,7 @@ export default {
                     range: ranges,
                     refinementList: refinementList,
                     hierarchicalMenu: { category_lvl1: categories.length ? categories : null },
-                    query: routeState.q,
+                    query: routeState.q || '__NO_QUERY__',
                     page: Number(routeState.page),
                     sortBy: routeState.sort,
                     hitsPerPage: Number(routeState.hits),