+++ resources/js/vue-components.js
@@ -59,6 +59,19 @@ Vue.component('autocomplete', () => ({
     },
     delay: 0,
 }))
+if (import.meta.env.VITE_DISABLE_DOUBLE_CLICK_FIX !== 'true') {
+    document.addEventListener('vue:loaded', function () {
+        // Workaround double click bug on ipad & iphone: https://stackoverflow.com/questions/71535540/keyboard-doesnt-open-in-ios-on-focus
+        if (
+            ['iPad Simulator', 'iPhone Simulator', 'iPod Simulator', 'iPad', 'iPhone', 'iPod'].includes(navigator.platform) ||
+            (navigator.userAgent.includes('Mac') && 'ontouchend' in document)
+        ) {
+            // Simply load the autocomplete ~600ms after pageload so it won't impact pageload but will
+            // Be fast enough that most users won't notice
+            setTimeout(() => window.document.dispatchEvent(new window.Event('loadAutoComplete')), 600)
+        }
+    })
+}
 Vue.component('checkout-login', () => import('./components/Checkout/CheckoutLogin.vue'))
 Vue.component('login', () => import('./components/User/Login.vue'))
 Vue.component('listing', () => import('./components/Listing/Listing.vue'))
@@ -66,3 +79,4 @@ Vue.component('search-suggestions', () => import('./components/Listing/SearchSug
 Vue.component('checkout-success', () => import('./components/Checkout/CheckoutSuccess.vue'))
 Vue.component('popup', () => import('./components/Popup.vue'))
 Vue.component('range-slider', () => import('./components/Elements/RangeSlider.vue'))
+Vue.component('recently-viewed', () => import('./components/RecentlyViewed.vue'))