+++ resources/views/cart/item/remove.blade.php
@@ -9,7 +9,7 @@
     <button
         v-on:click="mutate"
         title="@lang('Remove')"
-        class="hover:underline"
+        class="hover:underline text-red-700"
         :disabled="$root.loading"
         data-testid="cart-item-remove"
     >