+++ resources/views/checkout/pages/payment.blade.php
@@ -8,7 +8,7 @@
     <div class="container">
         @include('rapidez::checkout.partials.progressbar')
         <div v-if="hasCart" v-cloak>
-            <div class="flex gap-5 max-xl:flex-col">
+            <div class="flex gap-10 max-xl:flex-col">
                 <form class="w-full rounded bg h-fit p-4 xl:p-8 xl:w-3/4" v-on:submit.prevent="(e) => {
                         submitPartials(e.target?.form ?? e.target)
                             .then((result) =>
@@ -16,9 +16,9 @@
                                 && window.app.$emit('placeOrder')
                             ).catch();
                     }">
-                    @include('rapidez::checkout.steps.payment_method')
+                    @include('rapidez::checkout.steps.payment-method')
                     <div class="mt-3">@include('rapidez::checkout.steps.agreements')</div>
-                    @include('rapidez::checkout.steps.place_order')
+                    @include('rapidez::checkout.steps.place-order')
                 </form>
                 <div class="w-full xl:w-1/4">
                     @include('rapidez::checkout.partials.sidebar')