+++ resources/views/components/autocomplete/input.blade.php
@@ -8,7 +8,6 @@
 >
     <x-rapidez::input
         {{ $attributes->merge([
-            'id' => 'autocomplete-input',
             'type' => 'search',
             'name' => 'q',
             'autocomplete' => 'off',
@@ -18,6 +17,7 @@
             'placeholder' => __('What are you looking for?'),
             'class' => 'text-base h-12 peer',
             'data-testid' => 'autocomplete-input',
+            'id' => 'autocomplete-input',
         ]) }}
     />
     <button
@@ -30,9 +30,9 @@
         <x-heroicon-s-x-mark class="size-7" />
     </button>
     <x-rapidez::button
-        class="absolute right-0 top-0 bg-opacity-0 hover:bg-opacity-0 border-none *:peer-placeholder-shown:bg-muted *:peer-placeholder-shown:text"
+        class="!absolute right-0 top-0 bg-opacity-0 hover:bg-opacity-0 border-none *:peer-placeholder-shown:bg-muted *:peer-placeholder-shown:text"
         type="submit"
-        title="@lang('Search')"
+        :title="__('Search')"
     >
         <x-rapidez::autocomplete.magnifying-glass />
     </x-rapidez::button>