+++ resources/views/components/productlist.blade.php
@@ -1,6 +1,15 @@
 @props(['value', 'title' => false, 'field' => 'sku'])
+@slots(['items'])
 
-@if ($value)
+{{--
+Examples:
+<x-rapidez::productlist :value="['MS04', 'MS05', 'MS09']"/>
+<x-rapidez::productlist value="productIds" field="entity_id"/>
+<x-rapidez::productlist :value="false" filter-query-string="sku:MS04,MS05,MS09"/>
+<x-rapidez::productlist :value="false" v-bind:base-filters="() => [{dslQuery}}]"/>
+--}}
+
+@if ($value !== [])
     <lazy v-slot="{ intersected }">
         <listing
             {{ $attributes }}
@@ -15,21 +24,29 @@
                     :index-name="listingSlotProps.index"
                     :middlewares="listingSlotProps.middlewares"
                 >
-                    <ais-configure :filters="'{{ $field }}:({{ is_array($value)
-                        ? implode(' OR ', $value)
-                        : "'+".$value.".join(' OR ')+'"
-                    }})'"/>
+                    @slotdefault('before')
+                        @if ($value && $value !== [])
+                            <ais-configure :filters="'{{ $field }}:({{ is_array($value)
+                                ? implode(' OR ', $value)
+                                : "'+".$value.".join(' OR ')+'"
+                            }})'"/>
+                        @endif
+                    @endslotdefault
 
-                    <ais-hits v-slot="{ items, sendEvent }">
+                    <ais-hits v-slot="{ items, sendEvent }" v-bind:transform-items="listingSlotProps.transformItems">
                         <div v-if="items.length" class="flex flex-col gap-5">
                             @if ($title)
                                 <strong class="font-bold text-2xl">
                                     @lang($title)
                                 </strong>
                             @endif
-                            <x-rapidez::slider />
+                            @slotdefault('items')
+                                <x-rapidez::slider />
+                            @endslotdefault
                         </div>
                     </ais-hits>
+
+                    {{ $after ?? '' }}
                 </ais-instant-search>
             </div>
         </listing>