+++ tests/playwright/pages/BasePage.js
@@ -1,17 +1,36 @@
-import { expect } from '@playwright/test'
+import { test, expect } from '@playwright/test'
 import AxeBuilder from '@axe-core/playwright'
+import { playAudit } from 'playwright-lighthouse'
+import playwright from 'playwright'
+import lighthouseMobileConfig from 'lighthouse/core/config/lr-mobile-config.js'
 
 export class BasePage {
+    static tags = { tag: process.env.MAGENTO_VERSION ? '@' + process.env.MAGENTO_VERSION : null }
+
     constructor(page) {
         this.page = page
     }
 
-    async screenshot(type) {
-        let options = {}
+    async screenshot(type = '', options = {}) {
+        const masks = [this.page.getByTestId('masked')]
+        const emailFields = this.page.locator('[name=email]')
+
+        // Only mask filled email fields
+        const emailFieldsCount = await emailFields.count()
+        if (emailFieldsCount) {
+            for (let i = 0; i < emailFieldsCount; i++) {
+                let emailField = emailFields.nth(i)
+                if (await emailField.inputValue()) {
+                    masks.push(emailField)
+                }
+            }
+        }
+
+        options['mask'] = masks
 
         if (type.startsWith('fullpage')) {
             await this.scrolldown()
-            options = { fullPage: true }
+            options['fullPage'] = true
         }
 
         if (type.startsWith('fullpage-footer')) {
@@ -52,5 +71,62 @@ export class BasePage {
         })
 
         expect(accessibilityScanResults.violations).toEqual([])
+    }
+
+    async lighthouse(url) {
+        test.skip(test.info().project.name !== 'chromium', 'Chromium only')
+
+        const browser = await playwright['chromium'].launch({
+            args: ['--remote-debugging-port=9222'],
+        })
+        const page = await browser.newPage()
+        const reportName = `lighthouse-${new Date().getTime()}`
+
+        await page.goto(url)
+
+        try {
+            await playAudit({
+                page: page,
+                port: 9222,
+                thresholds: {
+                    performance: 90,
+                    accessibility: 100,
+                    'best-practices': 100,
+                    seo: 100,
+                },
+                reports: {
+                    formats: {
+                        html: true,
+                    },
+                    name: reportName,
+                },
+                config: {
+                    ...lighthouseMobileConfig,
+                    settings: {
+                        ...lighthouseMobileConfig.settings,
+                        skipAudits: [
+                            ...lighthouseMobileConfig.settings.skipAudits,
+                            // Skip everything that's not fixed within CI tests.
+                            'meta-description',
+                            'is-on-https',
+                            'redirects-http',
+                            'uses-long-cache-ttl',
+                            'uses-optimized-images',
+                            'cache-insight',
+                            'image-delivery-insight',
+                        ],
+                    },
+                },
+            })
+        } catch (error) {
+            await test.info().attach(reportName, {
+                path: 'lighthouse/' + reportName + '.html',
+                contentType: 'text/html',
+            })
+
+            throw error
+        } finally {
+            await browser.close()
+        }
     }
 }