+++ tests/playwright/search.spec.js
@@ -1,19 +1,22 @@
 import { test, expect } from '@playwright/test'
 import { ProductPage } from './pages/ProductPage'
+import { BasePage } from './pages/BasePage'
 
-test('search page', async ({ page }) => {
+test('search page', BasePage.tags, async ({ page }) => {
     const product = await new ProductPage(page).goto(process.env.PRODUCT_URL_SIMPLE)
     await page.goto('/search?q=' + product.name)
     await page.waitForLoadState('networkidle')
     await expect(page.getByTestId('listing-item').first()).toContainText(product.name)
-    await expect(page).toHaveScreenshot()
+    await new BasePage(page).screenshot()
 })
 
-test('autocomplete', async ({ page }) => {
+test('autocomplete', BasePage.tags, async ({ page }) => {
     const product = await new ProductPage(page).goto(process.env.PRODUCT_URL_SIMPLE)
     await page.goto('/')
+    await page.getByTestId('autocomplete-input').click()
+    await page.waitForLoadState('networkidle')
     await page.getByTestId('autocomplete-input').fill(product.name)
     await page.waitForLoadState('networkidle')
     await expect(page.getByTestId('autocomplete-item').first()).toContainText(product.name)
-    await expect(page).toHaveScreenshot()
+    await new BasePage(page).screenshot()
 })